requirements.txt
pip~=23.2.1
emoji~=2.14.1
attrs~=23.1.0
PySocks~=1.7.1
protobuf~=6.31.1
pillow~=10.4.0
pytest~=7.4.4
typing_extensions~=4.13.2
setuptools~=63.2.0
pluggy~=1.6.0
exceptiongroup~=1.1.3
iniconfig~=2.1.0
py~=1.11.0
numpy~=2.1.2
certifi~=2023.7.22
Brotli~=1.1.0
urllib3~=1.26.20
idna~=3.4
charset-normalizer~=3.4.2
packaging~=24.1
tomli~=2.2.1
platformdirs~=4.3.8
selenium~=4.15.2
PyYAML~=6.0.2
PyAutoGUI~=0.9.54
ddddocr~=1.5.6
requests~=2.32.3
allure-python-commons~=2.14.2